move MLP to model class, keep repr. generic
